Rename ARXPayloadAnalyser to ARXReIdentificationRiskFactory
Pull out duplicated code from ARXReIdentificationRiskFactory methods
